+++ calculation/sha_calc/sha_calc/directivity/bea20/directivity.py
@@ -1,5 +1,6 @@
 import math
 from pathlib import Path
+from typing import List
 
 import numpy as np
 import pandas as pd
@@ -9,10 +10,10 @@ from IM_calculation.source_site_dist import src_site_dist
 from sha_calc.directivity.bea20 import bea20, utils
 
 
-def compute_directivity_effects(
+def compute_directivity_srf_single(
     srf_file: str, srf_csv: Path, sites: np.ndarray, period: float = 3.0
 ):
-    """Computes directivity effects at the given sites with the given srf data
+    """Computes directivity effects at the given sites with the given srf data for a single hypocentre
 
     Parameters
     ----------
@@ -25,24 +26,195 @@ def compute_directivity_effects(
     period: float, optional
         Float to indicate which period to extract from fD to get fDi
     """
-    # Get rake, magnitude from srf_csv
-    mag = pd.read_csv(srf_csv)["magnitude"][0]
-    rake = pd.read_csv(srf_csv)["rake"][0]
+    (
+        mag,
+        rake,
+        planes,
+        lon_lat_depth,
+    ) = _srf_pre_process(srf_file, srf_csv)
 
-    # Get planes and points from srf_file
-    planes = srf.read_header(srf_file, idx=True)
-    points = srf.read_latlondepth(srf_file)
+    hyp_along_strike = 1
+    hyp_down_dip = 1
 
-    # Convert points to non dictionary format
-    lon_lat_depth = np.asarray([[x["lon"], x["lat"], x["depth"]] for x in points])
+    return compute_fault_directivity(
+        lon_lat_depth,
+        planes,
+        sites,
+        hyp_along_strike,
+        hyp_down_dip,
+        mag,
+        rake,
+        period,
+    )
 
+
+def compute_directivity_srf_multi(
+    srf_file: str, srf_csv: Path, sites: np.ndarray, period: float = 3.0
+):
+    """Computes directivity effects at the given sites with the given srf data using hypocentre averaging
+
+    Parameters
+    ----------
+    srf_file: str
+        String of the ffp to the location of the srf file
+    srf_csv: Path
+        Path to the location of the srf csv file
+    sites: np.ndarray
+        Numpy array full of site lon/lat values [[lon, lat],...]
+    period: float, optional
+        Float to indicate which period to extract from fD to get fDi
+    """
+    (
+        mag,
+        rake,
+        planes,
+        lon_lat_depth,
+    ) = _srf_pre_process(srf_file, srf_csv)
+
+    # TODO Update with best practice for hypocentre averaging
+    hyp_along_strike = 10
+    hyp_down_dip = 2
+
+    return compute_fault_directivity(
+        lon_lat_depth,
+        planes,
+        sites,
+        hyp_along_strike,
+        hyp_down_dip,
+        mag,
+        rake,
+        period,
+    )
+
+
+def compute_fault_directivity(
+    lon_lat_depth: np.ndarray,
+    planes: List,
+    sites: np.ndarray,
+    hyp_along_strike: int,
+    hyp_down_dip: int,
+    mag: float,
+    rake: float,
+    period: float,
+):
+    """
+    Does the computation of directivity for a fault with any number of hypocentres.
+    Can compute regardless if data came from an srf or nhm file.
+
+    Parameters
+    ----------
+    lon_lat_depth: np.ndarray
+        Each point of the srf fault in an array with the format [[lon, lat, depth],...]
+    planes: List
+        List of the planes that make up the fault
+    sites: np.ndarray
+        Numpy array full of site lon/lat values [[lon, lat],...]
+    hyp_along_strike: int
+        Number of hypocentres across strike
+    hyp_down_dip: int
+        Number of hypocentres down dip
+    mag: float
+        The magnitude of the fault
+    rake: float
+        The rake of the fault
+    period: float
+        The period for fdi to extract from the bea20 models fd
+    """
+    nominal_strike, nominal_strike2 = utils.calc_nominal_strike(lon_lat_depth)
+
+    if hyp_down_dip == 1 and hyp_along_strike == 1:
+
+        # Gets the plane index of the hypocentre
+        plane_index = 0
+        for index, plane in enumerate(planes):
+            if plane["dhyp"] != -999.99:
+                plane_index = index
+                break
+
+        return _compute_directivity_effect(
+            lon_lat_depth,
+            planes,
+            plane_index,
+            sites,
+            nominal_strike,
+            nominal_strike2,
+            mag,
+            rake,
+            period,
+        )
+    else:
+        # Customise the planes to set different hypocentres
+        n_hypo = hyp_down_dip * hyp_along_strike
+        depths = [i / (hyp_down_dip + 1) for i in range(1, hyp_down_dip + 1)]
+        planes_list, planes_index = utils.set_hypocentres(n_hypo, planes, depths)
+
+        # Creating the array to store all fdi values
+        fdi_array = []
+
+        for index, planes in enumerate(planes_list):
+            # Gets the plane index of the hypocentre
+            plane_index = planes_index[index]
+
+            fdi, _ = _compute_directivity_effect(
+                lon_lat_depth,
+                planes,
+                plane_index,
+                sites,
+                nominal_strike,
+                nominal_strike2,
+                mag,
+                rake,
+                period,
+            )
+
+            fdi_array.append(fdi)
+
+        fdi_average = np.mean(fdi_array, axis=0)
+
+        return fdi_average, fdi_array
+
+
+def _compute_directivity_effect(
+    lon_lat_depth: np.ndarray,
+    planes: List,
+    plane_index: int,
+    sites: np.ndarray,
+    nominal_strike: np.ndarray,
+    nominal_strike2: np.ndarray,
+    mag: float,
+    rake: float,
+    period: float,
+):
+    """
+    Does the computation of directivity and GC2 given a set of planes with a set hypocentre.
+
+    Parameters
+    ----------
+    lon_lat_depth: np.ndarray
+        Each point of the srf fault in an array with the format [[lon, lat, depth],...]
+    planes: List
+        List of the planes that make up the fault
+    plane_index: int
+        The index in planes that the hypocentre is located in
+    sites: np.ndarray
+        Numpy array full of site lon/lat values [[lon, lat],...]
+    nominal_strike: np.ndarray
+        The nominal strike coordinates (edge of the fault) with the highest longitiude value
+    nominal_strike2: np.ndarray
+        The nominal strike coordinates (edge of the fault) with the lowest longitiude value
+    mag: float
+        The magnitude of the fault
+    rake: float
+        The rake of the fault
+    period: float
+        The period for fdi to extract from the bea20 models fd
+    """
     # Calculate rx ry from GC2
     rx, ry = src_site_dist.calc_rx_ry_GC2(
         lon_lat_depth, planes, sites, hypocentre_origin=True
     )
 
     # Gets the s_max values from the two end points of the fault
-    nominal_strike, nominal_strike2 = utils.calc_nominal_strike(lon_lat_depth)
     rx_end, ry_end = src_site_dist.calc_rx_ry_GC2(
         lon_lat_depth, planes, nominal_strike, hypocentre_origin=True
     )
@@ -50,13 +222,6 @@ def compute_directivity_effects(
         lon_lat_depth, planes, nominal_strike2, hypocentre_origin=True
     )
     s_max = (min(ry_end, ry_end2)[0], max(ry_end, ry_end2)[0])
-
-    # Gets the plane index of the hypocentre
-    plane_index = 0
-    for index, plane in enumerate(planes):
-        if plane["dhyp"] == -999.99:
-            plane_index = index
-            break
 
     # Trig to calculate extra features of the fault for directivity based on plane info
     z_tor = planes[plane_index]["dtop"]
@@ -68,8 +233,31 @@ def compute_directivity_effects(
     d = (planes[plane_index]["dhyp"] - z_tor) / math.sin(math.radians(dip))
 
     # Use the bea20 model to work out directivity (fd) at the given sites
-    fd, fdi, phi_red, phi_redi, predictor_functions, other = bea20.bea20(
+    fd, fdi, phi_red, phi_redi, predictor_functions, other = bea20(
         mag, ry, rx, s_max, d, t_bot, d_bot, rake, dip, period
     )
 
-    return fd, fdi
+    return fdi, (fd, phi_red, phi_redi, predictor_functions, other)
+
+
+def _srf_pre_process(srf_file: str, srf_csv: Path):
+    """
+    Does the srf pre processing steps for computing directivity such as getting the
+    magnitude, rake, planes and lon_lat_depth values.
+
+    Parameters
+    ----------
+    srf_file: str
+        String of the ffp to the location of the srf file
+    srf_csv: Path
+        Path to the location of the srf csv file
+    """
+    # Get rake, magnitude from srf_csv
+    mag = pd.read_csv(srf_csv)["magnitude"][0]
+    rake = pd.read_csv(srf_csv)["rake"][0]
+
+    # Get planes and points from srf_file
+    planes = srf.read_header(srf_file, idx=True)
+    lon_lat_depth = srf.read_srf_points(srf_file)
+
+    return mag, rake, planes, lon_lat_depth